+++ adapters/utils/runAdapter.ts
@@ -465,8 +465,7 @@ async function _runAdapter({
   }
 
   async function setChainValidStart(chain: string) {
-    // Hourly backfills must admit the first hour, not wait until a full day has elapsed.
-    const cleanPreviousDayTimestamp = cleanCurrentDayTimestamp - WINDOW_SECONDS
+    const cleanPreviousDayTimestamp = cleanCurrentDayTimestamp - ONE_DAY_IN_SECONDS
     let _start = adapterObject![chain]?.start ?? 0
     // Use root-level deadFrom if set, otherwise use chain-specific deadFrom
     let _end = module.deadFrom ?? adapterObject![chain]?.deadFrom ?? 32503593600
